+++ api/server/httprest/server.go
@@ -44,7 +44,9 @@ func New(ctx context.Context, opts ...Option) (*Server, error) {
 		return nil, errors.New("router option not configured")
 	}
 	var handler http.Handler
-	if g.cfg.timeout > 10 {
+	defaultReadHeaderTimeout := time.Second
+	if g.cfg.timeout > 0*time.Second {
+		defaultReadHeaderTimeout = g.cfg.timeout
 		handler = http.TimeoutHandler(g.cfg.router, g.cfg.timeout, "request timed out")
 	} else {
 		handler = g.cfg.router
@@ -52,7 +54,7 @@ func New(ctx context.Context, opts ...Option) (*Server, error) {
 	g.server = &http.Server{
 		Addr:              g.cfg.httpAddr,
 		Handler:           handler,
-		ReadHeaderTimeout: time.Second,
+		ReadHeaderTimeout: defaultReadHeaderTimeout,
 	}
 
 	return g, nil